Refactor metacard revert logic to be reusable
