+++ ui-backend/catalog-ui-search/src/main/java/org/codice/ddf/catalog/ui/metacard/MetacardApplication.java
@@ -13,7 +13,6 @@
  */
 package org.codice.ddf.catalog.ui.metacard;
 
-import static ddf.catalog.util.impl.ResultIterable.resultIterable;
 import static javax.ws.rs.core.HttpHeaders.CONTENT_TYPE;
 import static javax.ws.rs.core.MediaType.APPLICATION_JSON;
 import static javax.ws.rs.core.MediaType.TEXT_PLAIN;
@@ -30,21 +29,12 @@ import static spark.Spark.post;
 import static spark.Spark.put;
 
 import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.ImmutableSet;
-import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
-import com.google.common.io.ByteSource;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.reflect.TypeToken;
 import ddf.catalog.CatalogFramework;
-import ddf.catalog.content.data.ContentItem;
-import ddf.catalog.content.data.impl.ContentItemImpl;
-import ddf.catalog.content.operation.impl.CreateStorageRequestImpl;
-import ddf.catalog.content.operation.impl.UpdateStorageRequestImpl;
-import ddf.catalog.core.versioning.DeletedMetacard;
 import ddf.catalog.core.versioning.MetacardVersion;
-import ddf.catalog.core.versioning.MetacardVersion.Action;
 import ddf.catalog.core.versioning.impl.MetacardVersionImpl;
 import ddf.catalog.data.Attribute;
 import ddf.catalog.data.AttributeDescriptor;
@@ -59,26 +49,15 @@ import ddf.catalog.data.impl.MetacardImpl;
 import ddf.catalog.data.impl.ResultImpl;
 import ddf.catalog.data.impl.types.SecurityAttributes;
 import ddf.catalog.data.types.Core;
-import ddf.catalog.federation.FederationException;
-import ddf.catalog.filter.FilterBuilder;
 import ddf.catalog.operation.DeleteResponse;
-import ddf.catalog.operation.QueryResponse;
-import ddf.catalog.operation.ResourceResponse;
 import ddf.catalog.operation.UpdateResponse;
 import ddf.catalog.operation.impl.CreateRequestImpl;
 import ddf.catalog.operation.impl.DeleteRequestImpl;
-import ddf.catalog.operation.impl.QueryImpl;
-import ddf.catalog.operation.impl.QueryRequestImpl;
-import ddf.catalog.operation.impl.ResourceRequestById;
 import ddf.catalog.operation.impl.SourceResponseImpl;
 import ddf.catalog.operation.impl.UpdateRequestImpl;
-import ddf.catalog.resource.ResourceNotFoundException;
-import ddf.catalog.resource.ResourceNotSupportedException;
 import ddf.catalog.source.IngestException;
 import ddf.catalog.source.SourceUnavailableException;
-import ddf.catalog.source.UnsupportedQueryException;
 import ddf.catalog.transform.QueryResponseTransformer;
-import ddf.catalog.util.impl.ResultIterable;
 import ddf.security.SubjectIdentity;
 import ddf.security.SubjectOperations;
 import ddf.security.audit.SecurityLogger;
@@ -102,10 +81,7 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
-import java.util.concurrent.Callable;
-import java.util.concurrent.TimeUnit;
 import java.util.function.Function;
-import java.util.function.Supplier;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import javax.ws.rs.NotFoundException;
@@ -118,7 +94,7 @@ import org.codice.ddf.catalog.ui.metacard.associations.Associated;
 import org.codice.ddf.catalog.ui.metacard.edit.AttributeChange;
 import org.codice.ddf.catalog.ui.metacard.edit.MetacardChanges;
 import org.codice.ddf.catalog.ui.metacard.history.HistoryResponse;
-import org.codice.ddf.catalog.ui.metacard.internal.OperationPropertySupplier;
+import org.codice.ddf.catalog.ui.metacard.history.MetacardRevertService;
 import org.codice.ddf.catalog.ui.metacard.notes.NoteConstants;
 import org.codice.ddf.catalog.ui.metacard.notes.NoteMetacard;
 import org.codice.ddf.catalog.ui.metacard.notes.NoteUtil;
@@ -135,8 +111,6 @@ import org.codice.ddf.catalog.ui.util.EndpointUtil;
 import org.codice.ddf.security.Security;
 import org.codice.gsonsupport.GsonTypeAdapters.DateLongFormatTypeAdapter;
 import org.codice.gsonsupport.GsonTypeAdapters.LongDoubleTypeAdapter;
-import org.opengis.filter.Filter;
-import org.opengis.filter.sort.SortBy;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import spark.servlet.SparkApplication;
@@ -146,12 +120,6 @@ public class MetacardApplication implements SparkApplication {
   private static final Logger LOGGER = LoggerFactory.getLogger(MetacardApplication.class);
 
   private static final String UPDATE_ERROR_MESSAGE = "Item is either restricted or not found.";
-
-  private static final Set<Action> CONTENT_ACTIONS =
-      ImmutableSet.of(Action.VERSIONED_CONTENT, Action.DELETED_CONTENT);
-
-  private static final Set<Action> DELETE_ACTIONS =
-      ImmutableSet.of(Action.DELETED, Action.DELETED_CONTENT);
 
   private static final String ERROR_RESPONSE_TYPE = "error";
 
@@ -167,8 +135,6 @@ public class MetacardApplication implements SparkApplication {
   private static final Type ASSOCIATED_EDGE_LIST_TYPE =
       new TypeToken<List<Associated.Edge>>() {}.getType();
 
-  private static final int PAGE_SIZE = 250;
-
   private static final Gson GSON =
       new GsonBuilder()
           .disableHtmlEscaping()
@@ -178,7 +144,6 @@ public class MetacardApplication implements SparkApplication {
           .create();
 
   private final CatalogFramework catalogFramework;
-  private final FilterBuilder filterBuilder;
   private final EndpointUtil util;
   private final Validator validator;
   private final WorkspaceTransformer transformer;
@@ -199,15 +164,14 @@ public class MetacardApplication implements SparkApplication {
 
   private final AssociatedQueryMetacardsHandler queryMetacardsHandler;
 
-  private OperationPropertySupplier operationPropertySupplier;
-
   private SubjectOperations subjectOperations;
 
   private SecurityLogger securityLogger;
 
+  private MetacardRevertService metacardRevertService;
+
   public MetacardApplication(
       CatalogFramework catalogFramework,
-      FilterBuilder filterBuilder,
       EndpointUtil endpointUtil,
       Validator validator,
       WorkspaceTransformer transformer,
@@ -222,9 +186,9 @@ public class MetacardApplication implements SparkApplication {
       SubjectIdentity subjectIdentity,
       WorkspaceService workspaceService,
       AssociatedQueryMetacardsHandler queryMetacardsHandler,
-      Security security) {
+      Security security,
+      MetacardRevertService metacardRevertService) {
     this.catalogFramework = catalogFramework;
-    this.filterBuilder = filterBuilder;
     this.util = endpointUtil;
     this.validator = validator;
     this.transformer = transformer;
@@ -239,14 +203,7 @@ public class MetacardApplication implements SparkApplication {
     this.subjectIdentity = subjectIdentity;
     this.workspaceService = workspaceService;
     this.queryMetacardsHandler = queryMetacardsHandler;
-  }
-
-  public void addOperationPropertySupplier(OperationPropertySupplier operationPropertySupplier) {
-    this.operationPropertySupplier = operationPropertySupplier;
-  }
-
-  public void removeOperationPropertySupplier(OperationPropertySupplier operationPropertySupplier) {
-    this.operationPropertySupplier = null;
+    this.metacardRevertService = metacardRevertService;
   }
 
   private String getSubjectEmail() {
@@ -383,7 +340,7 @@ public class MetacardApplication implements SparkApplication {
         (req, res) -> {
           String id = req.params(":id");
           String storeId = req.params(":storeId");
-          List<Result> queryResponse = getMetacardHistory(id, storeId);
+          List<Result> queryResponse = metacardRevertService.getMetacardHistory(id, storeId);
           if (queryResponse.isEmpty()) {
             res.status(204);
             return "[]";
@@ -409,7 +366,7 @@ public class MetacardApplication implements SparkApplication {
           String id = req.params(":id");
           String revertId = req.params(":revertid");
           String storeId = req.params(":storeId");
-          Metacard versionMetacard = revert(id, revertId, storeId);
+          Metacard versionMetacard = metacardRevertService.revert(id, revertId, storeId);
           return util.metacardToJson(MetacardVersionImpl.toMetacard(versionMetacard));
         });
 
@@ -868,61 +825,6 @@ public class MetacardApplication implements SparkApplication {
     exception(RuntimeException.class, util::handleRuntimeException);
   }
 
-  private Metacard revert(String id, String revertId, String storeId)
-      throws UnsupportedQueryException, SourceUnavailableException, FederationException,
-          IngestException, ResourceNotFoundException, IOException, ResourceNotSupportedException {
-    try {
-      Map<String, Serializable> properties =
-          operationPropertySupplier == null
-              ? new HashMap<>()
-              : operationPropertySupplier.properties(OperationPropertySupplier.QUERY_TYPE);
-
-      Metacard versionMetacard = util.getMetacardById(revertId, properties);
-
-      List<Result> queryResponse = getMetacardHistory(id, storeId);
-      if (queryResponse.isEmpty()) {
-        throw new NotFoundException("Could not find metacard with id: " + id);
-      }
-
-      Optional<Metacard> contentVersion =
-          queryResponse
-              .stream()
-              .map(Result::getMetacard)
-              .filter(
-                  mc ->
-                      getVersionedOnDate(mc).isAfter(getVersionedOnDate(versionMetacard))
-                          || getVersionedOnDate(mc).equals(getVersionedOnDate(versionMetacard)))
-              .filter(mc -> CONTENT_ACTIONS.contains(Action.ofMetacard(mc)))
-              .filter(mc -> mc.getResourceURI() != null)
-              .filter(mc -> ContentItem.CONTENT_SCHEME.equals(mc.getResourceURI().getScheme()))
-              .min(
-                  Comparator.comparing(
-                      mc ->
-                          util.parseToDate(
-                              mc.getAttribute(MetacardVersion.VERSIONED_ON).getValue())));
-
-      if (!contentVersion.isPresent()) {
-        /* no content versions, just restore metacard */
-        revertMetacard(versionMetacard, id, false);
-      } else {
-        revertContentandMetacard(contentVersion.get(), versionMetacard, id);
-      }
-
-      securityLogger.audit("Reverted {} to metacard {}", revertId, id);
-
-      return versionMetacard;
-    } catch (UnsupportedQueryException
-        | SourceUnavailableException
-        | FederationException
-        | IngestException
-        | ResourceNotFoundException
-        | IOException
-        | ResourceNotSupportedException e) {
-      securityLogger.audit("Failed to revert {} to metcard {}", revertId, id, e);
-      throw e;
-    }
-  }
-
   private Set<String> getHiddenFields(List<Result> metacards) {
     Set<String> matchedHiddenFields;
     List<Pattern> hiddenFieldPatterns =
@@ -947,150 +849,6 @@ public class MetacardApplication implements SparkApplication {
                         .anyMatch(pattern -> pattern.test(attr)))
             .collect(Collectors.toSet());
     return matchedHiddenFields;
-  }
-
-  private void revertMetacard(Metacard versionMetacard, String id, boolean alreadyCreated)
-      throws SourceUnavailableException, IngestException {
-    LOGGER.trace("Reverting metacard [{}] to version [{}]", id, versionMetacard.getId());
-    Metacard revertMetacard = MetacardVersionImpl.toMetacard(versionMetacard);
-    Action action =
-        Action.fromKey((String) versionMetacard.getAttribute(MetacardVersion.ACTION).getValue());
-
-    if (DELETE_ACTIONS.contains(action)) {
-      attemptDeleteDeletedMetacard(id);
-      if (!alreadyCreated) {
-        Map<String, Serializable> properties =
-            operationPropertySupplier == null
-                ? new HashMap<>()
-                : operationPropertySupplier.properties(OperationPropertySupplier.CREATE_TYPE);
-
-        catalogFramework.create(
-            new CreateRequestImpl(Collections.singletonList(revertMetacard), properties));
-      }
-    } else {
-      tryUpdate(
-          4,
-          () -> {
-            catalogFramework.update(new UpdateRequestImpl(id, revertMetacard));
-            return true;
-          });
-    }
-  }
-
-  private void revertContentandMetacard(Metacard latestContent, Metacard versionMetacard, String id)
-      throws SourceUnavailableException, IngestException, ResourceNotFoundException, IOException,
-          ResourceNotSupportedException {
-    LOGGER.trace(
-        "Reverting content and metacard for metacard [{}]. \nLatest content: [{}] \nVersion metacard: [{}]",
-        id,
-        latestContent.getId(),
-        versionMetacard.getId());
-    Map<String, Serializable> properties = new HashMap<>();
-    properties.put("no-default-tags", true);
-    ResourceResponse latestResource =
-        catalogFramework.getLocalResource(
-            new ResourceRequestById(latestContent.getId(), properties));
-
-    ContentItemImpl contentItem =
-        new ContentItemImpl(
-            id,
-            new ByteSourceWrapper(() -> latestResource.getResource().getInputStream()),
-            latestResource.getResource().getMimeTypeValue(),
-            latestResource.getResource().getName(),
-            latestResource.getResource().getSize(),
-            MetacardVersionImpl.toMetacard(versionMetacard));
-
-    // Try to delete the "deleted metacard" marker first.
-    boolean alreadyCreated = false;
-    Action action =
-        Action.fromKey((String) versionMetacard.getAttribute(MetacardVersion.ACTION).getValue());
-    if (DELETE_ACTIONS.contains(action)) {
-      alreadyCreated = true;
-      catalogFramework.create(
-          new CreateStorageRequestImpl(
-              Collections.singletonList(contentItem), id, new HashMap<>()));
-    } else {
-      // Currently we can't guarantee the metacard will exist yet because of the 1
-      // second
-      // soft commit in solr. this busy wait loop should be fixed when alternate
-      // solution
-      // is found.
-      tryUpdate(
-          4,
-          () -> {
-            catalogFramework.update(
-                new UpdateStorageRequestImpl(
-                    Collections.singletonList(contentItem), id, new HashMap<>()));
-            return true;
-          });
-    }
-    LOGGER.trace("Successfully reverted metacard content for [{}]", id);
-    revertMetacard(versionMetacard, id, alreadyCreated);
-  }
-
-  private void trySleep() {
-    try {
-      Thread.sleep(350);
-    } catch (InterruptedException e) {
-      Thread.currentThread().interrupt();
-    }
-  }
-
-  private void tryUpdate(int retries, Callable<Boolean> func) throws IngestException {
-    if (retries <= 0) {
-      throw new IngestException("Could not update metacard!");
-    }
-    LOGGER.trace("Trying to update metacard.");
-    try {
-      func.call();
-      LOGGER.trace("Successfully updated metacard.");
-    } catch (Exception e) {
-      LOGGER.trace("Failed to update metacard");
-      trySleep();
-      tryUpdate(retries - 1, func);
-    }
-  }
-
-  private void attemptDeleteDeletedMetacard(String id) {
-    LOGGER.trace("Attemping to delete metacard [{}]", id);
-    Filter tags =
-        filterBuilder.attribute(Metacard.TAGS).is().like().text(DeletedMetacard.DELETED_TAG);
-    Filter deletion =
-        filterBuilder.attribute(DeletedMetacard.DELETION_OF_ID).is().equalTo().text(id);
-    Filter filter = filterBuilder.allOf(tags, deletion);
-
-    Map<String, Serializable> properties =
-        operationPropertySupplier == null
-            ? new HashMap<>()
-            : operationPropertySupplier.properties(OperationPropertySupplier.QUERY_TYPE);
-
-    QueryResponse response = null;
-    try {
-      response =
-          catalogFramework.query(
-              new QueryRequestImpl(new QueryImpl(filter), false, null, properties));
-    } catch (UnsupportedQueryException | SourceUnavailableException | FederationException e) {
-      LOGGER.debug("Could not find the deleted metacard marker to delete", e);
-    }
-
-    if (response == null || response.getResults() == null || response.getResults().size() != 1) {
-      LOGGER.debug("There should have been one deleted metacard marker");
-      return;
-    }
-
-    final DeleteRequestImpl deleteRequest =
-        new DeleteRequestImpl(response.getResults().get(0).getMetacard().getId());
-    deleteRequest.getProperties().put("operation.query-tags", ImmutableSet.of("*"));
-    try {
-      catalogFramework.delete(deleteRequest);
-    } catch (IngestException | SourceUnavailableException e) {
-      LOGGER.debug("Could not delete the deleted metacard marker", e);
-    }
-    LOGGER.trace("Deleted delete marker metacard successfully");
-  }
-
-  private Instant getVersionedOnDate(Metacard mc) {
-    return util.parseToDate(mc.getAttribute(MetacardVersion.VERSIONED_ON).getValue());
   }
 
   private AttributeDescriptor getDescriptor(Metacard target, String attribute) {
@@ -1164,36 +922,6 @@ public class MetacardApplication implements SparkApplication {
         .equals(AttributeType.AttributeFormat.DATE);
   }
 
-  private List<Result> getMetacardHistory(String id, String sourceId) {
-    Map<String, Serializable> properties =
-        operationPropertySupplier == null
-            ? new HashMap<>()
-            : operationPropertySupplier.properties(OperationPropertySupplier.QUERY_TYPE);
-
-    Filter historyFilter =
-        filterBuilder.attribute(Metacard.TAGS).is().equalTo().text(MetacardVersion.VERSION_TAG);
-    Filter idFilter =
-        filterBuilder.attribute(MetacardVersion.VERSION_OF_ID).is().equalTo().text(id);
-
-    Filter filter = filterBuilder.allOf(historyFilter, idFilter);
-    Set<String> sources = sourceId != null ? Collections.singleton(sourceId) : null;
-    ResultIterable resultIterable =
-        resultIterable(
-            catalogFramework,
-            new QueryRequestImpl(
-                new QueryImpl(
-                    filter,
-                    1,
-                    PAGE_SIZE,
-                    SortBy.NATURAL_ORDER,
-                    false,
-                    TimeUnit.SECONDS.toMillis(10)),
-                false,
-                sources,
-                properties));
-    return Lists.newArrayList(resultIterable);
-  }
-
   private Metacard updateMetacard(String id, Metacard metacard)
       throws SourceUnavailableException, IngestException {
     return catalogFramework
@@ -1220,18 +948,5 @@ public class MetacardApplication implements SparkApplication {
 
   public void setSecurityLogger(SecurityLogger securityLogger) {
     this.securityLogger = securityLogger;
-  }
-
-  private static class ByteSourceWrapper extends ByteSource {
-    Supplier<InputStream> supplier;
-
-    ByteSourceWrapper(Supplier<InputStream> supplier) {
-      this.supplier = supplier;
-    }
-
-    @Override
-    public InputStream openStream() {
-      return supplier.get();
-    }
   }
 }
