+++ ui-backend/catalog-ui-search/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -255,14 +255,19 @@ Implementation details
                availability="optional"
                filter="(type=revert)">
         <reference-listener
-                ref="metacardApplication"
+                ref="metacardRevertService"
                 bind-method="addOperationPropertySupplier"
                 unbind-method="removeOperationPropertySupplier"/>
     </reference>
 
-    <bean id="metacardApplication" class="org.codice.ddf.catalog.ui.metacard.MetacardApplication">
+    <bean id="metacardRevertService" class="org.codice.ddf.catalog.ui.metacard.history.MetacardRevertService">
         <argument ref="catalogFramework"/>
         <argument ref="filterBuilder"/>
+        <property name="securityLogger" ref="securityLogger" />
+    </bean>
+
+    <bean id="metacardApplication" class="org.codice.ddf.catalog.ui.metacard.MetacardApplication">
+        <argument ref="catalogFramework"/>
         <argument ref="endpointUtil"/>
         <argument ref="validator"/>
         <argument ref="workspaceTransformer"/>
@@ -278,6 +283,7 @@ Implementation details
         <argument ref="queryWorkspaceService"/>
         <argument ref="queryMetacardsHandler"/>
         <argument ref="ddfSecurity" />
+        <argument ref="metacardRevertService" />
         <property name="subjectOperations" ref="subjectOperations1" />
         <property name="securityLogger" ref="securityLogger" />
     </bean>
